// src/app/Models/SeizureManAnimalConflict.ts
export interface SeizureManAnimalConflict {
    id: number;
    sNo: number;
    nameOfGammaUnit: string;
    placeOfOccurrence: string;
    noOfFPFPersonnelDeployed: number;
    remarks: string;
    provinceId: number;
    districtId: number;
    month: number;
    year: number;
    lastUpdatedOn: Date;
    isActive: boolean;
}